+++ lib/laflem/collections/python/__init__.py
@@ -7,6 +7,7 @@ from .virtualenv import (
     VirtualenvCleanupModule,
     VirtualenvCreateModule,
     VirtualenvListModule,
+    VirtualenvDeleteModule,
 )
 
 
@@ -20,5 +21,6 @@ class PythonCollection(BaseCollection):
     modules = {
         "venv-create": VirtualenvCreateModule,
         "venv-list": VirtualenvListModule,
+        "venv-delete": VirtualenvDeleteModule,
         "venv-cleanup": VirtualenvCleanupModule,
     }